Account creation completed
Complete account creation is finished, added some rules to the database to improve security

Images get uploaded to the firebase storage under /users/*userid*/*the image*
All images get cropped to a maximum height or width of 440 pixels, athey get compressed to a quality factor of 0.4. The result of this is a blob object which gets converted to a base64 object. The api used for this is compressionjs & documentation is available at: https://github.com/fengyuanchen/compressorjs/blob/master/README.md

user object in /models/ had some tweaks, might need a few more as the project progresses

Most important edits are in the following files:
UserService.js
testenv.js